chore(utils): merge snip9.ts with outsideExecution.ts
