+++ src/utils/outsideExecution.ts
@@ -2,6 +2,9 @@ import { CallData } from './calldata';
 import { getSelectorFromName } from './selector';
 import { encodeShortString } from './shortString';
 import { Call } from '../types/lib';
+import { Calldata, Signature } from '../types';
+import { formatSignature } from './stark';
+
 import {
   OutsideExecutionOptions,
   OutsideCall,
@@ -98,4 +101,17 @@ function callToTypedData(call: Call, version: EOutsideExecutionVersion) {
     Selector: outsideCall.selector,
     Calldata: outsideCall.calldata,
   };
+}
+
+// Builds a CallData for the execute_from_outside() entrypoint.
+export function buildExecuteFromOutsideCallData(
+  outsideExecution: OutsideExecution,
+  signature: Signature
+): Calldata {
+  const abiData = outsideExecution.getABIData();
+  const formattedSignature = formatSignature(signature);
+  return CallData.compile({
+    outside_execution: abiData, // TODO we should encode it field by field most likely
+    signature: formattedSignature,
+  });
 }
